fix metrics to ns

+++ core/src/main/java/org/example/metrics/MetricBucket.java
@@ -4,6 +4,6 @@ import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.atomic.AtomicInteger;
 
 public class MetricBucket {
-    public final ConcurrentLinkedQueue<Double> timings = new ConcurrentLinkedQueue<>();
+    public final ConcurrentLinkedQueue<Long> timings = new ConcurrentLinkedQueue<>();
     public final AtomicInteger count = new AtomicInteger(0);
 }
